Remove must_not from the random queries
This should be added back at some point, but right now it seems that
hyperion doesn't properly evaluate these conditions. For example, if a
query only has a must_not block, 0 hits are returned even when there
should be some hits (and running it in the UI as SPL returns some hits).